Unit testability: split out definition of statuses struct.
